Mimic minigzip behavior and only unlink files if not using -c copy out argument.

+++ test/minideflate.c
@@ -346,7 +346,7 @@ int main(int argc, char **argv) {
 
     if (fin != stdin) {
         fclose(fin);
-        if (!keep) {
+        if (!copyout && !keep) {
             unlink(argv[i]);
         }
     }
